Refactor: File paths, unused files
- Updated imports for all changes
- Removed unused files
- CreateUser deleted
- CreateUserForm refactored to UsersForm
- EditUser refactored to EditUserForm
- UserCard refactored to UserCard inside user-management
- UsersTable refactored to UsersList
- Deleted context providers
- Refactored admin dashboard and user management to same names inside admin
