+++ client/src/App.jsx
@@ -1,16 +1,12 @@
 import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
 import Layout from "./layouts/Layout";
 import Home from "./pages/home/Home";
-import Login from "./pages-new/login/Login";
+import Login from "./pages/login/Login";
 import NoPage from "./pages/no-page/NoPage";
-import AdminDashboard from "./pages-new/admin/AdminDashboard";
-import CustomRoute from "./components/custom-route/CustomRoute";
-import UserManagement from "./pages-new/admin/UserManagement";
-import { AuthProvider } from "./context/AuthProvider";
-import { UserManagementProvider } from "./context/UserManagementProvider";
-import { AlertProvider } from "./context/AlertProvider";
+import AdminDashboard from "./pages/admin/AdminDashboard";
+import UserManagement from "./pages/admin/UserManagement";
 import "./App.css";
-import Onboarding from "./pages-new/onboarding/Onboarding";
+import Onboarding from "./pages/onboarding/Onboarding";
 import PrivateRoute from "./routes/PrivateRoute";
 import AdminRoute from "./routes/AdminRoute";
 import PublicRoute from "./routes/PublicRoute";
