--- client/src/components/admin/user-management/CreateUser.jsx
@@ -1,31 +1,0 @@
-import React, { useState } from "react";
-import CreateUserForm from "./CreateUserForm";
-import { useUserManagement } from "../../../context/UserManagementProvider";
-import { useAlertOld } from "../../../context/AlertProvider";
-
-const CreateUser = () => {
-  const [createUserError, setCreateUserError] = useState("");
-  const { createUser, refreshUsers } = useUserManagement();
-  const { addAlertOld } = useAlertOld();
-
-  const handleCreateUser = async (userData) => {
-    try {
-      await createUser(userData);
-      addAlertOld("User Created Successfully", "success");
-      await refreshUsers();
-    } catch (error) {
-      addAlertOld(error.message, "danger");
-    }
-  };
-
-  return (
-    <>
-      <CreateUserForm
-        onSubmit={handleCreateUser}
-        errorMessage={createUserError}
-      />
-    </>
-  );
-};
-
-export default CreateUser;
