+++ client/src/components/common/SearchBox.jsx
